+++ app/src/main/java/de/htwBerlin/ai/mediAlarm/data/rhythm/TimePoint.kt
@@ -2,5 +2,5 @@ package de.htwBerlin.ai.mediAlarm.data.rhythm
 
 data class TimePoint(
     val type: TimepointType,
-    val absoluteTimeFromMidnight: Int?
+    val absoluteTimeFromMidnight: Long?
 )
